Set root to 400M and reserve remaining space
Carving out space for the dm-verity hashes combined with setting the
root disk size to the maximum available space is probably a bad idea for
the future.

This reserves 87M per half for future partition expansion.

To make the sgdisk command line more readable, sectors were replaced
with MiB sizes and the partitions are listed backwards so that we don't
have to calculate the last sector of each partition.

Signed-off-by: iliana destroyer of worlds <iweller@amazon.com>