simplify architecture handling and dependency generators
Functional aarch64 builds are pending fixes for the image build, so
the current multi-architecture builds are not adding any value.

In any case, this was largely a convenience for initial development,
allowing us to easily exercise package builds for two architectures.
For production builds, we can simply invoke `make` as many times as
needed to build for all architectures.

Take advantage of the required changes to the dependency generators
to clean up and streamline the code.

Signed-off-by: Ben Cressey <bcressey@amazon.com>