always build makedep files and enforce proper ordering
Although the included files are rebuilt if they do not exist, they
are not built in the listed order. Furthermore, the rules to add
dependencies to an existing target do not take effect until the
next time the Makefile is loaded.

The heavy-handed approach here always generates a makedep file,
even if it is empty, and does a one-off build of the crate to ensure
that it exists before the other scripts run.

Errors when building these files now also cause `make` to fail, and
avoid creating the output file. This will help avoid "successful"
runs that are doomed or broken in subtle ways.

Signed-off-by: Ben Cressey <bcressey@amazon.com>