Ensure we return consistent types in each branch of an API path
By saving the result in a typed Rust variable instead of serialized JSON, the
Rust compiler will tell us if branches return different types.

This exposed a &str versus String discrepancy, where &str was used for
efficiency in one case but couldn't be used in another; this change means we
have to copy strings in the one case, but it's better to be consistent.

Signed-off-by: Tom Kirchner <tjk@amazon.com>