Add DataStore.get_prefix to handle common retrieval pattern
More and more places need to retrieve sets of keys starting with some prefix.
This change implements it as a default method for DataStore.

It's then easy for users to turn the resulting HashMap into a structure using
from_map, or modify the keys as necessary, but those tasks are less common.

Signed-off-by: Tom Kirchner <tjk@amazon.com>