Clarify internal requirements of data in Key
I tried to add KeyType to Key and found the issue described in this comment
about the Hash value of a Key.  This documents the issue so we can more easily
make a decision if we want to do this in the future.

To help clarify, this makes the traits on Key and KeyType more accurate;
there's no reason to Ord-er Keys currently, and KeyType is safe to Eq/Hash.

An updated log message also helps debug this problem if we run into it again.

Signed-off-by: Tom Kirchner <tjk@amazon.com>