Simplify MemoryDataStore's commit method
We already have the pending data stored separately so we don't need get_prefix,
which also means we don't need to create new Key structures.

The fact that we only committed "settings" was copied from FilesystemDataStore,
and doesn't matter for MemoryDataStore, which is only used for testing.  No
need to subset commits here.

Signed-off-by: Tom Kirchner <tjk@amazon.com>