Refactor key name regular expressions to allow reuse
We want the ability to match individual segments of key names, and to allow
other modules to build new regexes using a consistent definition of a segment.
This adds a public segment definition, and refactors the regex definitions
around it.

Now that we have regex components and don't need the explanations in a (?x)
verbose regex, we can actually show them in an error message, so we no longer
need comments about coordinating changes.

Signed-off-by: Tom Kirchner <tjk@amazon.com>