nix: defend against "./sources" differences
A few of the newer packages have a sources file with minor differences
that throw off the source detection. To avoid this, the input is
"canonicalized" to single whitespace separated fields. Assertions are
added to validate generated data when parsing sources at generation
time and also use in the fetcher.