Add 'migrator' that runs data store migrations
This tool finds migrations on disk, runs them on a copy of the data store, and
if successful, atomically updates the data store links to flip to the new
version.

You tell it the data store to migrate (from which it finds the current version)
and the version to migrate to, and it understands the available migrations and
which are necessary to move to the requested version.

Note: this still assumes there's a single, fixed MIGRATIONS_PATH.  We don't yet
have the systems and locations we need to find migrations from the image and
migrations that have been downloaded, so this temporarily makes this
simplifying assumption.

Signed-off-by: Tom Kirchner <tjk@amazon.com>