release: add FQDN helper service
In the absence of a hostname override, kube-proxy reads its node name
from /proc/sys/kernel/hostname and expects the API server to recognize
it.

However, kubelet registers the node with the host's FQDN, so we need
to make sure they match. networkd, hostnamed, and resolved have all
the required data, but by design they avoid setting the hostname to a
value that might be specific to one of many network links.

For the EKS use case, there is only one network link that concerns us,
so we can write the value we want and sidestep the design discussion.

Signed-off-by: Ben Cressey <bcressey@amazon.com>