Allow metadata on unpopulated data keys
In 8d6dd1d0b8bf we added listing of metadata keys, and added the restriction
that metadata could only refer to populated data keys.  This was only an
implementation detail -- there was no design reason to exclude metadata for
unpopulated data keys -- and the limitation only extended as far as the
metadata list API.

We now have a use case for metadata on unpopulated data keys: metadata that
specifies how to generate the data key.

This change allows listing metadata for unpopulated data keys.  We do this by
walking the tree, the same way we do when listing data keys.  The old
implementation was already walking the tree to find data keys, and was
additionally globbing, so this should be faster and simpler.

To implement this, we add a KeyType structure to datastore::filesystem that
helps centralize the logic for creating Keys from Paths, and abstracts over
data and metadata keys during a list operation, so we can use the same
filesystem walk code.

Signed-off-by: Tom Kirchner <tjk@amazon.com>