Add modeled base64 type
The ValidBase64 type can be used in the model for input that we want to
validate is base64.  If it's invalid, it won't deserialize, so it won't be
accepted by the API.

The resulting value acts like a string (via Deref and friends) and represents
the input base64 text; its purpose is input validation, not decoding.  When
serialized back out, for example for a GET /settings, the original input is
returned, for symmetry and understandability.

The encoded value can be used later on, for example to populate a template, by
base64-decoding the string.

Signed-off-by: Tom Kirchner <tjk@amazon.com>