Build amazon-ssm-agent
This builds the SSM agent and installs some default conf and service
files into the image. A few SSM patches are included to work around some
build issues found in the current SSM repo when building for Linux.

Build steps are similar to the .spec file SSM normally uses for
packaging but includes debug information and a GNU-style build id to
cooperate with Thar's build system.

The two additional patches work around issues with launching the shell
for the "Session" module; allowing a root user and specifying an
absolute path to 'sh'.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>