sdk: enable pie and ssp by default
PIE is complicated to wire up through compiler and linker flags, since
whether or not it can be used depends on the output. It is simpler to
have the compiler enable it where possible.

SSP is easy to set up through compiler flags, but it's just as easy to
enable it by default.

Signed-off-by: Ben Cressey <bcressey@amazon.com>