Add 'configured' target
This represents the point in the boot process when all user-requested and
dynamic configuration has been applied to the system.  Services that depend on
input from user data, or from sundog's dynamic setting generators, can express
that as a systemd dependency to be sure they're complete and their
configuration is written.

Note: if you use a Requires dependency on configured.target, your service will
go down if (for example) moondog is restarted, which is probably not what you
want.  You can use a Wants dependency to be safe, but this means your service
will still start without configuration if moondog/sundog fail.

Signed-off-by: Tom Kirchner <tjk@amazon.com>