Implements 'latest known time' when checking for expirations
Everytime when the system time is sampled, its checked against the last
known system time. If the sampled system time is somehow earlier than the latest known time,  then
the system time is irrational and an error is raised. The newly sampled system time is serialized and stored as latest_known_time.json.