Use Unix-domain sockets for API server
This switches to actix-web for the HTTP server component, to get Unix-domain
socket support.

The only major interface change is that a PATCH to /settings should no longer
include the outer '{settings: {}}' layer, just the inner values; this was done
to improve consistency within the API.  It was much easier to implement that
way with actix, so it was done within this change.