ssm: Run in AL2 container
Run the SSM agent and tools inside a container instead of directly on
the host. For now this is built from an AL2 container with the SSM
resources added into the image. Since these files are now only used in
the container they are moved out of the normal file system hierarchy.

The API socket is mapped in and host networking enabled to give the SSM
agent access to the API server, but otherwise it has no access to the
rest of the system.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>