Only run moondog once, at first boot
Moondog writes settings from user data to the API server.  If the user changes
these settings after boot, and then they reboot, they wouldn't expect them to
be overwritten by their original user data settings.

This changes moondog to write to a marker file after successfully running, and
adds a check to its unit file that prevents it from running again if the marker
exists.  The user can override this by removing the file.

Eventually we will probably want user data that applies at different times,
with some of the flexibility of cloud-init, but this is a simple solution for
now that can be easily changed.

When we do need more flexibility, we'll likely want to store the relevant
moondog state in the data store through the API, but we don't yet have
application-level storage or APIs.