Add migrator service
The migrator service initiates migration of the data store to the version
specified in the image we just booted; this may be the same version, in which
case it will do nothing.

We want to run very early in boot so other components can rely on the fact that
the data store format version is what they expect.  In particular, it runs
before storewolf, because there's no need to migrate data that storewolf just
populated; storewolf comes from the new image and understands the new format.