updog: Teaching new tricks
These changes move Updog towards handling updates and update metadata as
described in the Thar documentation. In particular this starts to handle
update waves, update timing, and proper versioning based on the current
understanding of the metadata format.

Updates will be staged into "waves" which will be specified in the
update manifest. A wave is described by a "bound" number and a UTC time.
A series of bound-time points defines a simple update curve. The Updog
client will generate a random seed value which from then on is used to
locate itself in one of these waves, and a random amount of jitter is
added to the exact update time. When every wave has passed all clients
should be at the new image version.

Applicable updates filtered by version, flavor, and state. Version is
assumed to be of the form "vx.y.z", flavor is some string sourced
from the system, and states should reflect those described by TUF.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>