Finish initial host container integration
This commit adds the final pieces to integrate host containers. It adds
default settings for admin and control container metadata, configuration
files, and restart commands. The template files for also included in
the packaging so they end up on the OS's filesystem in the proper place.

This also removes the ssm package from the build, but not from the repo.