release: set up persistent storage for /var and /opt
This moves the THAR-DATA mount from /var/lib/thar to /local, where
we create the directories necessary to mount /var and /opt from
that volume.

Since the directories are persistent, we no longer need to create
the correspoindg tmpfs or overlayfs mounts during preinit.

Both mounts are now wanted by the local-fs.target, so most units do
not need to declare an explicit dependency.

Signed-off-by: Ben Cressey <bcressey@amazon.com>