tough_schema: Work around duplicate `_type` issue
When serializing a role, the `_type` tag gets included as expected. When
deserializing it, `_type` is not ignored and instead added to `_extra`,
meaning that when it's serialized and deserialized again, it fails due
to a duplicate key.

This works around the issue by removing the `_type` field from `_extra`
if present during deserialization.

Signed-off-by: iliana destroyer of worlds <iweller@amazon.com>