use conventional paths for container services
Although these services ship as part of the OS, we want to enable
customers to replace the persistent storage volume with their own
copy, for example to supply a cache of downloaded container images.

Paths such as `/var/lib/docker` make it easier for these volumes
to be created without baking in a lot of OS specific details.

Signed-off-by: Ben Cressey <bcressey@amazon.com>