tough: Permit missing length/hashes in snapshot.json
The TUF specification is inconsistent about the `METAFILES` struct. In
its first example, snapshot.json, it contains only a `version` field,
but in its second example, timestamp.json, it additionally contains the
`length` and `hashes` fields, while also saying "METAFILES is the same
is described for the snapshot.json file".

Previously we took the more strict definition, but now that I'm trying
to improve tough's interoperability in the TUF ecosystem, I'm reversing
that decision.

Signed-off-by: iliana destroyer of worlds <iweller@amazon.com>