updog: Return update jitter to caller
Instead of calculating some jitter and waiting, Updog will return the
calculated update time to the caller so that it can be called again at a
later time. The caller should then supply the start time to Updog when
calling it again.
If an update exists but Updog is in a later wave, the start of that wave
is returned.

This also adds a helper function `output()` to clean up a few spots
where Updog's output depends on whether --json has been set.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>