Allow user-defined host containers
This adds a host-containers service that asks the API for the defined
host-containers and tells systemd to start/stop them to match their 'enabled'
setting.  The settings are written to an EnvironmentFile that's read by the new
host-container@ service, a systemd templated service used for all host
containers by way of a suffix like "host-container@admin".

The existing container-specific thar-be-settings templated systemd unit files
were removed in favor of this more general approach.  Existing metadata for
these was also removed, replaced with a single metadata entry that invokes the
new tool.

The existing host-containers package, which just contains the Go host-ctr
source, was renamed to host-ctr.