updog: Logical fixes to update waves
Change Update.update_wave() to return the wave bounds that encompass
Updog's seed value instead of just the lower bound. This makes the
actual update 'wave' relevant to a given Updog easier to conceptualise.
With follow-on changes to update_ready() and jitter() this fixes an
issue where if Updog was in the "0th" wave (ie. no lower bound) it would
incorrectly assume it had to wait for the upper bound time, when it is
actually free to update.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>