updog: Don't source migrations from a downloaded root image
Updog downloads the root image before writing it to disk so that it can
be mounted and searched for existing migrations, in an effort to reduce
network traffic. However this has two problems in the current approach:
- It doesn't solve the issue of trust for the "other" side when it boots
and finds these migrations.
- If run from a container with smaller cgroup limits (eg. dogswatch)
it can invoke the OOM killer!

Be more conservative with memory and remove this approach. Future
patches will introduce a more trusted and less memory-hungry solution.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>