INSTALL: simplify documentation for subnet selection
Uses more advanced aws jmespath/jq queries to bring together information with
less work from the user, and with clearer output.

The subnet commands were in the middle of the cluster settings, too, which are
now moved above the subnet commands, so the user doesn't have to keep track of
those two sections mentally.