signpost: Avoid updating into nothing
Add two new commands to signpost:
- mark-inactive-valid, used to mark the inactive partition as having a
potentially valid image but not for boot.
- cancel-upgrade, which reverts the changes of upgrade-to-inactive.

In addition upgrade-to-inactive makes two initial checks; whether the
inactive partition has been marked valid with mark-inactive-valid, and
whether the inactive partition has already been marked for boot. The
first is to avoid accidentally booting into a blank partition, while the
second is a warning that makes it more clear that invoking
upgrade-to-inactive twice does not revert the change.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>