sundog: use JSON output to allow non-String generated model types
We couldn't use an integer for the bork-generated update seed because sundog
was taking the output of generators and serializing it as a string.  This lets
the generator determine the type of the output by using JSON as an interchange
format.

Pluto and bork are updated to serialize their output appropriately.

(Note: bork can't use a u64 because JSON can't encode a u64; u32 is granular
 enough for this purpose.)