kernel: Add kernel-devel package
When building the kernel also create a 'kernel-devel' RPM which includes
the configured kernel source tree and extra artifacts to support
building out-of-tree modules. This is not included in the Thar image,
but is instead intended to be downloaded later on a running instance via
a TUF target, for example.

Based off a similar script provided by Ben Cressey <bcressey@amazon.com>

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>