kernel: Enable access to /proc/config.gz
Tools such as sysdig[0] have the ability to deploy BPF probes across a
Kubernetes cluster. However part of the setup scripts for this require
access to the kernel config and on Thar this isn't available in any of
the common locations. CONFIG_IKCONFIG makes the config available
regardless of where the setup is being run and is a well known source,
so switch it on for Thar.

[0]: https://github.com/draios/sysdig

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>