[storewolf] Improve default metadata representation
This commit improves how metadata is represented in our defaults.toml.
Previously each metadata entry required a seperate 4 line declaration,
which is noisy and unpleasant. With this change, metadata for each
corresponding setting can be defined together in a single metadata
entry.

The commit adds a single function to translate this new defaults toml
format into a Vec<Metadata> which is processed in storewolf exactly
the same as we did previously.