build: adjust cache volume logic
Although the cache volume was previously created and was functional,
the subsequent build stage could not see the expected marker file
under either Docker 18.09 and 19.03.

With this change, the marker file is now visible, which verifies that
we are using the cache volume that matches the build stage.

This is a follow-up to the fix in f5c863c8.

Signed-off-by: Ben Cressey <bcressey@amazon.com>