Remove hardcoded paths from repo canary
We can use a tempdir for the data store, since we don't use it persistently.

We can accept an argument for the trusted root path; local usage would probably
point to the root.json in the same directory, and container-based canary usage
can point to the root file in the directory it creates.