build: standardize on variant nomenclature
Previously we referred to custom OS builds as "images" or "flavors".
Images as a term isn't accurate - we produce multiple filesystem and
disk images for each build. Flavors is better, but is arguably a
project-specific term for a more general concept.

This introduces the term "variant" for our custom builds, which aligns
with the corresponding field in the os-release file, and fixes up the
build logic to match.

Signed-off-by: Ben Cressey <bcressey@amazon.com>