grub: core.img: Reboot if configfile fails
If GRUB attempts to load a config file from a boot partition that is
empty or otherwise broken, it will sit forever at a rescue console,
which is not particularly helpful in EC2.

gptprio.next decrements tries_left on partitions not yet marked
successful, so rebooting will usually result in the correct behavior of
rolling back.

Some diagnostic output is printed before sleeping for 30 seconds to help
aid future debugging. No additional output is printed if the configfile
boots successfully.

Signed-off-by: iliana destroyer of worlds <iweller@amazon.com>