Use separate containerd configs in aws-k8s and aws-dev
We were failing to write the containerd config in the dev image because
pod-infra-container-image wasn't set -- because it lives in the Kubernetes
settings we don't have in the dev image.

We don't need a pod-infra-container-image (or similar) setting in the dev image
because we use containerd through Docker which doesn't use it.  So, the
clearest change is to split the containerd config file, removing the irrelevant
cri-plugin settings from the dev version, including the
pod-infra-container-image setting.