build: generate manifest with checksum on build
Including a manifest and the RELEASE.toml used permits checks made in
other tools that will drive the test and release automation.

The generated manifest needs to include followed symlinks: the thar
images that are linked to, when uploaded (if they are anywhere in the
process), become regular files copied from their linked targets in the
set of resulting artifacts.

Signed-off-by: Jacob Vallejo <jakeev@amazon.com>