Add "schnauzer" generator for parameterized setting values
This lets settings be generated based on a template that reference other
settings.  The first example is regional container image URIs, where we replace
part with the value of settings.aws.region.

The template is stored as a new metadata key, next to the setting-generator
metadata, which schnauzer will read from the API.  For example, if the template
is "foo-{{ settings.bar }}" and "settings.bar" is set to "baz", then the
setting output will be "foo-baz".

To be able to reference other settings from sundog-spawned generators, we need
to commit settings that have already been patched, e.g. from user data.  To do
this, this commit eliminates the settings-committer service, instead using the
settings-committer binary as an ExecStartPre on any service that requires
committed settings.