migrator: allow removing settings in migrations
Migrations weren't able to remove settings because we started with a copy of
the existing data store and wrote out any changes key-by-key, meaning a
"removal" would just leave the old copied key in place on disk.

With this change, we no longer copy the data store, we instead give migrations
the source data store and the target data store, and (via migration-helpers)
the input data is read from the source instead of a copy.  This is more
efficient because it avoids a needless copy, and allows the migration to
determine exactly what's left after it runs.