Add Rust tool for updating/signing TUF repository via pipeline
Adds a tool meant to be run in a CI environment that will update an
existing TUF repo with contents from a recent build. It updates all
of the metadata/target files, signs them and the new targets, and
places them in a directory suitable for syncing to your existing
TUF repository.