updog: only remove known extensions from migration filenames
If we always set_extension, it removes anything after a period.  If the
migration does not happen to be named with an extension, then the only period
is in the version number, so we change "migrate_0.1_foo" to "migrate_0", which
is not a valid migration filename.  The migration won't run and the upgrade
will fail.

With this change, we only remove the extension (.lz4) that matches our
compression format.

Migrations downloaded by older updog could not have an extension because it
would not be removed and would again be invalid, so unfortunately we have a
diverging path.  Versions before the addition of set_extension (let's call them
version A) cannot have an extension, and versions with set_extension but
without this change (let's call them version B) must have an extension, which
means that update manifests need to list special copies of migrations that have
an extension for any paths leading in/out of that intermediate version B.  This
must be maintained until version B is no longer supported.

If we call releases including this change "version C", then here's an example
set of updates to work around this issue:
```
  # Original migration before finding the issue; don't want to block update into B
  (A, B): ["migrate_0.B_foo"]
  # If starting from A and updating past B, skip over the special migration needed by B
  (A, C): ["migrate_0.B_foo", "migrate_0.C_bar"]
  # Special migration out of affected version B
  (B, C): ["migrate_0.C_bar.lz4"]
```

As long as B is supported, each new version will need a special path from B.
For example, after releasing version D:
```
  # Original migration before finding the issue; don't want to block update into B
  (A, B): ["migrate_0.B_foo"]
  # If starting from A and updating past B, skip over the special migration needed by B
  (A, C): ["migrate_0.B_foo", "migrate_0.C_bar"]
  # Normal new migrations go here, for new versions like D
  (C, D): ["migrate_0.D_baz"]
  # Special migration out of affected version B
  (B, C): ["migrate_0.C_bar.lz4"]
  # We need a path from B to each new version so they have extensions to be removed
  (B, D): ["migrate_0.C_bar.lz4", "migrate_0.D_baz.lz4"]
```