updog: Add a standalone messenger dog to send query parameters
Add the 'smoky'[0] utility to send query parameters outside of the
normal update process, and without otherwise interacting with the TUF
repository.

Smoky uses the HttpQueryTransport already used by Updog to add query
parameters to some URL, and then invokes the default fetch method. The
URL must be provided since Smoky is likely to be called before such
information is available.

Since this doesn't require the tough repository it can be used on boot
to mark boot success or failure, without depending on the rest of the
system being available.

[0]: https://en.wikipedia.org/wiki/Smoky_(dog)

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>