migration-helpers: back-compat: remove datastore keys removed in migration
The old migrator interface used a single data store as source and target.
After a migration changes the MigrationData, the resulting values are written
out to the data store.  Since this data store also served as the source, it had
all of the old keys on disk already, and nothing explicitly removed them; we'd
only see new keys overwrite old.

To allow downgrading to versions with the old migrator interface, we need
migrations to explicitly remove keys from the data store.  This change to
migration-helpers detects keys that were removed from MigrationData by a
migration and tells the data store to explicitly remove them.  As long as
migrations are rebuilt against this library, they'll work in the old or new
migrator.

This workaround can be removed when we no longer support versions with the old
migrator interface.