refresh-timestamp-lambda: adds `threshold` parameter
Adds a new threshold parameter that defines a timeframe that makes the
lambda push out the updated and signed timestamp.json to the TUF repo. If the current
date hasn't pass the threshold date, the lambda only signs but does not
push to TUF repository.

Moves the signing of timestamp.json into a separate function.