Update RELEASE.toml for next release
This adds the remove-region migration needed to downgrade after adding
platform-specific settings.

This updates the migration syntax to match manifest.json from the TUF repo,
both for simplicity of updating the manifest and because we have a lot of
information and the increased density makes it easier for humans to see the
various migration paths.