infra: Stacks for S3 log processing and access
Define two Cloudformation stacks, one intended for the account owning
the TUF repository, and the other for an account intended for telemetry
purposes.

The first stack, "update-logs-glue-role", defines resources to allow an
AWS Glue job to process raw S3 access logs into a more consumable
format, and give the telemetry account access to the results. In
particular this is intended to use the Glue job defined by the scripts
in https://github.com/awslabs/athena-glue-service-logs.

The second stack, "update-logs-cross-account-athena-role", defines a
role which has access to the resulting bucket, and permissions required
for the Glue and Athena services to operate on it. It then creates a
Glue crawler which will on-demand parse the resulting database from the
first stack, and make it available to Athena it its account.

Signed-off-by: Samuel Mendoza-Jonas <samjonas@amazon.com>