build libaudit
This is a minimal build of audit that gives us `auditctl` and the
libraries it needs to parse and apply rules.

In place of `auditd`, we have a service that loads a set of rules
so that only SELinux events are logged. This configuration can be
replaced by a suitably privileged container if necessary.

Signed-off-by: Ben Cressey <bcressey@amazon.com>