release: add dependency on audit
If the audit subsystem is not explicitly disabled with "audit=0" on
the kernel command line, it will be enabled by journald but left in
an unconfigured state. Various events such as iptables changes will
then be logged, and the performance cost can be significant.

Our audit package contains a service to load audit rules at boot.

Signed-off-by: Ben Cressey <bcressey@amazon.com>