controller: simplify exit condition
This is a simplified and equivalent exit condition for the
controller.Run function.  The previous code was as follows:

for {
	select {
	case <-ctx.Done():
		cancel()
		return nil
	}
}

That is an infinite loop, selecting on the Done channel whenever it has
content or is closed.  The Done channel is a context channel, and will
be closed when the parent context (ctx) is canceled.  Since this is an
infinite loop with only one channel in the select, we can simplify that
to perform a regular, blocking read on the channel:

<-ctx.Done()

The other portion of this code is the body, which consists of calling
cancel().  The cancel function here is the CancelFunc of a derived
context called worker.  When a context is derived, it is automatically
canceled when the parent context is canceled.  Thus if the ctx.Done()
channel is ready (has content or is closed), we know already that the
child context (worker) is canceled and there is no need to call
cancel().