Rename workspaces directory to 'sources', package to 'os'
workspaces was originally a set of Cargo workspaces, but we now build them
together because it's much faster to build.  This renames the 'workspaces'
directory to the clearer 'sources' because it contains our source code.  The
'workspaces' package is renamed to 'os' to represent that it contains our
top-level / userspace OS components.