Remove update update-checking boot service and timer
This was added during early development to give us a signal about available
updates, with a sense that it might have helped with automatic updates.  We've
since refined our thought, building the update operator for the aws-k8s
variant, and starting to plan better solutions for other variants.

Also, this causes a failed service (and therefore a "degraded" system) when
there are no updates, when in fact having no updates is usually a healthy
state.