Move migrations from small private partition to data partition
The private partition is very tight on space, and we could easily reach a point
where we weren't able to download migrations and accomplish an update.  The
private partition is used for the data store, so there's also contention
between those two critical functions.  Storing migrations on the larger data
partition takes a bit of space away from the user, but seems a reasonable
compromise to allow them to upgrade.

Rather than update the path in os.spec, we remove it, because it inaccurately
reflected the idea that we control the directory at build time, when in fact
it's created and populated at runtime.