common_migrations: Remove HashMap hack using local struct
This was found when adding a top-level Model struct, which the old comment said
would help with this case, but it actually didn't because we can't use the
model in migrations.  (Would you use the old or the new?)

Instead, we can remove the HashMap by making a single-use struct with a
'settings' key.  This makes it more obvious that only "settings" keys are
supported, so make the docstring reflect that reality too.