systemd: lower timeout during system shutdown
If we are still running processes during shutdown, they are likely to
be running in containers rather than managed by the host system. They
may not expect or respond to SIGTERM, which can delay the restart for
up to 90 seconds.

In the case of an update, we expect containers to be drained by the
operator before the system is restarted. However, if the system is
powered off directly then this may not happen.

With the network down, it is unlikely that processes can complete any
useful work apart from syncing data to disk. A lower timeout means we
will reboot or power off more quickly, which allows the node or its
replacement to come up faster.

Signed-off-by: Ben Cressey <bcressey@amazon.com>