Revert "containerd: specify killmode=mixed"
This reverts commit 75125f672a1ac8ebd2dc82a9e206e2ad50c9af63.

The reason for changing KillMode was to make the system shut down
more quickly. However, this is flawed in practice because although
the containerd shims are killed more quickly, the processes running
inside the containers are not since they are no longer part of the
unit's control group.

This configuration is also discouraged by upstream, as it means that
the containerd service cannot be safely restarted.