release: lower start and stop timeouts
The default timeout for start and stop is 90 seconds, and none of the
current host services require this much time. If they did, we could
override the setting locally in the service unit.

Many of the processes on the system end up running inside scopes that
are dynamically created by the orchestrator agent. By changing the
default timeout, we ensure that these processes are stopped quickly
during shutdown or restart.

Signed-off-by: Ben Cressey <bcressey@amazon.com>