selinux-policy: add all classes and actions to policy
This extends the initial implementation of the policy with a complete
set of all classes and available actions as of the 5.4 kernel. We also
include the permissions expected by systemd and dbus-broker. In a few
cases, actions that no longer exist in the kernel have been removed,
for example the "entrypoint" permission for the "chr_file" class.

The more comprehensive policy now denies unknown actions by default,
which creates some risk for kernel updates, but also acts as a nudge
to revisit the policy periodically. Any newly added classes and
actions will be logged by the kernel when the policy is loaded.

The 5.4 kernel added support for restricting file watches through
inotify(), dnotify(), and fanotify(). Related rules were added as a
special case to the policy to prevent processes running in containers
from blocking the host's access to its own files. Other uses are
still allowed.

Since classes related to sockets, networks, and IPCs are now defined
in the policy, it also includes rules to permit almost all actions
without restriction. The only exceptions are the actions related to
SELinux labeling, all of which are reserved for the host OS.

Because the policy defines userspace classes and actions, we can now
restrict interactions with systemd and dbus-broker to the host OS.
These programs are regarded as implementation details and not meant
for direct use by processes running in containers.

Other management functions, notably including the ability to load and
manage the SELinux policy, have also been reserved for the host OS.

We now include all relevant capabilities in the policy, to enable the
full set of policy features, and to ensure that all defined classes
and actions are used.

Signed-off-by: Ben Cressey <bcressey@amazon.com>