Add MIT/Apache license file hashes to backtrace-sys clarify statement
These license files were added in the crates/backtrace-sys directory of the
archive, as symlinks to the files in the top level of the archive.  This makes
the scanner find them for the backtrace-sys crate explicitly, so they need to
be listed in the license files of the backtrace-sys clarification, otherwise we
get a file mismatch error.