systemd: relocate more files to console package
Features related to passwords and emergency debugging assume that an
interactive shell will be present for use on the console. Variants
that do not include a shell also do not need the related files.

The new sulogin helper enables the single-user login shell to work as
expected for development and debugging purposes.

Signed-off-by: Ben Cressey <bcressey@amazon.com>