Add variant override to updog arguments
If --variant is given, updog will use its argument to look for updates, rather
than the currently running variant.  This allows for manually changing between
variants that you know are compatible.

Implementation note: running_version() was removed so that the current variant
isn't as easily used by accident when the user intended to override it.