selinux-policy: protect snapshot layers from modification
Now that we have the machinery in place to provide a "context" option
for overlayfs mounts, we can use a read-only type for file objects in
the lower directories. We also record the type in `lxc_contexts` for
programs that rely on that file for discovery.

`host-ctr` now runs with the same type as the other container runtime
components, because it handles the `mount()` syscall. The kernel will
try to create a "work" subdirectory in workdir during the mount, so
the calling process needs permissions to do so.

Signed-off-by: Ben Cressey <bcressey@amazon.com>