selinux-policy: confine system daemons and protect state
Before saving clock measurements and DHCP leases to durable storage,
we want to make sure they can't be modified by an untrusted workload.

chronyd and wicked are both network-facing daemons, so we have the
additional goal of limiting their access to the host OS, to mitigate
the impact of a compromise.

dbus-broker does not need to persist any state, but could be exposed
to untrusted local input, and benefits from similar confinement.

systemd also persists some state, such as the journal and the random
seed, so we label that data as well to prevent unexpected changes.

Signed-off-by: Ben Cressey <bcressey@amazon.com>