systemd: generate stable machine ID for Xen guests
We don't have a good way to persist the machine ID in a place that's
accessible during early boot, so our best bet is to always generate
the same ID.

Both Xen and KVM can provide a UUID to identify guests, and systemd
already uses the KVM value if available. This patch extends the logic
to check for the Xen value as well.

Signed-off-by: Ben Cressey <bcressey@amazon.com>