docker: privileged containers should have SELinux
By default, Docker injects the equivalent of --security-opt
label:disable for containers launched with the --privileged option with
the intent of providing as much privilege as possible.  This option
disables both process and mount labeling for the container.
Unfortunately, disabling labeling does not provide additional privileges
when Bottlerocket's SELinux policy is enforced but instead reduces
privileges.  This change removes the injection of the label:disable
option so privileged containers receive the same default labels as
non-privileged containers.

To launch a container with all SELinux labeling disabled, the explicit
--security-opt label:disable can still be used.
To launch a container with a label that provides additional privileges,
--security-opt label:type:super_t can be used.

Fixes: https://github.com/bottlerocket-os/bottlerocket/issues/1011