Fix warnings for use of deprecated items in common_migrations
The singular `AddSettingMigration` and `RemoveSettingMigration` exist for
backward compatibility and are marked deprecated.  The trait implementations on
those items have to reference the deprecated items, of course, but this
triggers compiler warnings about use of the deprecated items.  We only want
warnings when people are actively using these types outside of the module that
defines them; we don't want to see the warnings on every build.